extract CPK type test helpers
